Chips/Hisilicon: Add IoInitDxe driver
IoInitDxe driver is to initialize high speed link (SERDES) and SMMU
both for UEFI and OS.

SMMU configuration code is imported from PV660 SMMU driver for OS
(internal initial version); we move the configuration to UEFI for
PV660 SMMU is not totally compatible with ARM SMMUv3 and the code
can't be upstreamed to Linux kernel. Therefore, part of the code does
not conform to EDK2 code style.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Heyi Guo <heyi.guo@linaro.org>