D02/D03/PCIe: delete PNP0C02 configuration
This patch is to solve the below bug. "PNP0C02" attched two modules
drives/pnp/system.c and drivers/net/fjes/fjes_main.c.
The module of "fjes"  led to the call trace.
The temporary solution is to delete the "PNP0C02" in uefi to solve the bug.
We think it is a bug that "fjes" driver use id "PNP0C02". We have sent
a email to netdev@vger.kernel.org for flaging this.

[    4.232071] Call trace:
[    4.234526] [<ffffffc000613ec8>] driver_probe_device+0x318/0x478
[    4.240555] [<ffffffc0006140d4>] __driver_attach+0xac/0xb0
[    4.246062] [<ffffffc000611794>] bus_for_each_dev+0x80/0xd4
[    4.251655] [<ffffffc0006136a8>] driver_attach+0x2c/0x34
[    4.256987] [<ffffffc000613184>] bus_add_driver+0x210/0x298
[    4.262579] [<ffffffc000614cf8>] driver_register+0x68/0xfc
[    4.268085] [<ffffffc0004bccf0>] acpi_bus_register_driver+0x48/0x54
[    4.274381] [<ffffffbffc009044>] fjes_init_module+0x44/0x1000 [fjes]
[    4.280758] [<ffffffc000082afc>] do_one_initcall+0xfc/0x1ec
[    4.286352] [<ffffffc0001c3940>] do_init_module+0x6c/0x1d0
[    4.291859] [<ffffffc00014aaac>] load_module+0x1b14/0x233c
[    4.297365] [<ffffffc00014b53c>] SyS_finit_module+0xc4/0xf4
[    4.302959] [<ffffffc000085c70>] el0_svc_naked+0x24/0x28

https://bugs.linaro.org/show_bug.cgi?id=2292

Signed-off-by: Dongdong Liu <liudongdong3@huawei.com>